Keep flow config tests in weighted shard
